# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 1OF1 Trader Pro Affiliate Back Office API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/applications:
    get:
      operationId: getAdminApplications
      tags: [admin]
      summary: Get all applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, denied]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationsListResponse"
  /admin/applications/{id}:
    get:
      operationId: getAdminApplication
      tags: [admin]
      summary: Get application detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffiliateApplication"
  /admin/applications/{id}/approve:
    post:
      operationId: approveApplication
      tags: [admin]
      summary: Approve application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationActionRequest"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/applications/{id}/deny:
    post:
      operationId: denyApplication
      tags: [admin]
      summary: Deny application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationActionRequest"
      responses:
        "200":
          description: Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Admin - Affiliates
...
  /admin/affiliates/{id}/commission:
    patch:
      operationId: updateAffiliateCommission
      tags: [admin]
      summary: Update affiliate commission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCommissionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/commissions:
    get:
      operationId: getAdminCommissions
      tags: [admin]
      summary: Get all commissions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, paid, voided]
        - name: affiliateId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Commissions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommissionsListResponse"
  /admin/commissions/{id}:
    patch:
      operationId: adjustCommission
      tags: [admin]
      summary: Manually adjust a commission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustCommissionRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Admin - Sales/Memberships
...
  /affiliate/commissions:
    get:
      operationId: getAffiliateCommissions
      tags: [affiliate]
      summary: Get affiliate's commissions
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Commissions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AffiliateCommissionsListResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    AffiliateApplication:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        country:
          type: string
        telegram:
          type: string
          nullable: true
        discord:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        twitterUrl:
          type: string
          nullable: true
        youtubeUrl:
          type: string
          nullable: true
        audienceType:
          type: string
        communitySize:
          type: string
        trafficSources:
          type: string
        whyJoin:
          type: string
        tradingExperience:
          type: string
        payoutMethod:
          type: string
          nullable: true
        payoutDetails:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, approved, denied]
        reviewedBy:
          type: integer
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - phone
        - country
        - audienceType
        - communitySize
        - trafficSources
        - whyJoin
        - tradingExperience
        - status
        - createdAt
    ApplicationActionRequest:
      type: object
      properties:
        notes:
          type: string
          nullable: true
    ApplicationsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AffiliateApplication"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...
    UpdateCommissionRequest:
      type: object
      properties:
        defaultCommissionType:
          type: string
          enum: [flat, percent]
        defaultCommissionValue:
          type: number
      required:
        - defaultCommissionType
        - defaultCommissionValue
...
    Commission:
      type: object
      properties:
        id:
          type: integer
        affiliateId:
          type: integer
        affiliateName:
          type: string
          nullable: true
        paymentId:
          type: integer
          nullable: true
        customerId:
          type: integer
          nullable: true
        customerEmail:
          type: string
          nullable: true
        commissionType:
          type: string
          enum: [flat, percent]
        commissionValue:
          type: number
        commissionAmount:
          type: number
        commissionStatus:
          type: string
          enum: [pending, approved, paid, voided]
        reason:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - affiliateId
        - commissionType
        - commissionValue
        - commissionAmount
        - commissionStatus
        - createdAt
    CommissionsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Commission"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPending:
          type: number
        totalApproved:
          type: number
      required:
        - data
        - total
        - page
        - limit
        - totalPending
        - totalApproved
    AdjustCommissionRequest:
      type: object
      properties:
        commissionAmount:
          type: number
          nullable: true
        commissionStatus:
          type: string
          enum: [pending, approved, paid, voided]
          nullable: true
        reason:
          type: string
          nullable: true
...
    AffiliateCommissionRow:
      type: object
      properties:
        id:
          type: integer
        customerEmail:
          type: string
          nullable: true
        customerName:
          type: string
          nullable: true
        saleAmount:
          type: number
          nullable: true
        amount:
          type: number
        commissionType:
          type: string
        status:
          type: string
        rateType:
          type: string
        rateValue:
          type: number
        createdAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - customerEmail
        - customerName
        - saleAmount
        - amount
        - commissionType
        - status
        - rateType
        - rateValue
        - createdAt
        - paidAt
    AffiliateCommissionsListResponse:
      type: object
      properties:
        commissions:
          type: array
          items:
            $ref: "#/components/schemas/AffiliateCommissionRow"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalEarned:
          type: number
        totalPending:
          type: number
        totalPaid:
          type: number
      required:
        - commissions
        - total
        - page
        - limit
        - totalEarned
        - totalPending
        - totalPaid
...